build: redefine uint64_t and int64_t size
It's really annoying that on 64bit both of them are defined as `long`
and not the more logical `long long`. This doesn't break anything
per-se, however it generates warnings for using printf %llx over %lx,
which would not work in 32bit code (where it is indeed defined to be
`long long`).