fix: invalid invlpg implementation
Basically two problems: First, the inline assembly syntax was meant for
AT&T and not Intel syntax. Thus the idea was (translated into Intel)
`invlpg [virtual_address]`. here comes the second issue, the constraint
is "m", aka memory, but the `()` in the asm already achieve the idea and
were probably meant for "r" constrain (which is used in linux btw).