fix: not setting user_supervisor for every page map entry
Similarly to `read_write` bit, user_supervisor is `&`-ed to to get with
all the paging map entries to get the final bit. Which I don't
completely see the usage for, it just makes the control less granular
(although we can "disable" it), as we do here.
One use I do see, is for example making the kernel pml3 and pml2
supervisor, hence preventing accidental maps (which shouldn't occur in
the first place). Still weird.