fix: prevent arp cache mutex deadlock
Drops a packet if the mutex is locked. The only other way to resolve it
(which is not implementing kernel threads) is to not acknowledge that we
read the packet (and wait for the next interrupt), but this might
eventually just cause other packets dropped.

Another solution would be to implement a queue of entries to add, and
add them right before the mutex is unlocked.